# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/udpateBracket":{"post":{"tags":["bracket-rest-controller"],"operationId":"updateBracket","responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOListListListUser"}}}}}}},"/checkEventStatus/{eventName}":{"get":{"tags":["bracket-rest-controller"],"operationId":"checkStatus","parameters":[{"name":"eventName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOBoolean"}}}}}}},"/getEventBracket/{eventName}":{"get":{"tags":["bracket-rest-controller"],"operationId":"getBracket","parameters":[{"name":"eventName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOListListListUser"}}}}}}},"/getTopEvents/{maxAmount}":{"get":{"tags":["event-rest-controller"],"operationId":"getTopEvents","parameters":[{"name":"token","in":"query","required":true,"schema":{"$ref":"#/components/schemas/JwtAuthenticationToken"}},{"name":"maxAmount","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOListTournament"}}}}}}},"/findEventByName":{"get":{"tags":["event-rest-controller"],"operationId":"findEventListing","parameters":[{"name":"token","in":"query","required":true,"schema":{"$ref":"#/components/schemas/JwtAuthenticationToken"}},{"name":"eventName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOTournament"}}}}}}},"/findEventsAttended":{"get":{"tags":["event-rest-controller"],"operationId":"findAttendedEvents","parameters":[{"name":"token","in":"query","required":true,"schema":{"$ref":"#/components/schemas/JwtAuthenticationToken"}}],"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOListTournament"}}}}}}},"/findEventsCreated":{"get":{"tags":["event-rest-controller"],"operationId":"findEventsCreated","parameters":[{"name":"token","in":"query","required":true,"schema":{"$ref":"#/components/schemas/JwtAuthenticationToken"}}],"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOListTournament"}}}}}}},"/createEvent":{"post":{"tags":["event-rest-controller"],"operationId":"createEvent","parameters":[{"name":"title","in":"query","required":true,"schema":{"type":"string"}},{"name":"description","in":"query","required":true,"schema":{"type":"string"}},{"name":"gameTitle","in":"query","required":true,"schema":{"type":"string"}},{"name":"selectedStartDate","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"capacity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"token":{"$ref":"#/components/schemas/JwtAuthenticationToken"},"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOString"}}}}}}},"/findEvents":{"get":{"tags":["event-rest-controller"],"operationId":"findEvents","parameters":[{"name":"eventName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOListTournament"}}}}}}},"/leaderboard":{"get":{"tags":["leaderboard-rest-controller"],"operationId":"leaderboardStats","parameters":[{"name":"token","in":"query","required":true,"schema":{"$ref":"#/components/schemas/JwtAuthenticationToken"}}],"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOListUser"}}}}}}},"/storeUser":{"post":{"tags":["user-rest-controller"],"operationId":"storeUserInfo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationToken"}}}},"responses":{"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DTOString"}}}}}}}},"components":{"schemas":{"DTOListListListUser":{"type":"object","properties":{"statusCode":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"type":"array","items":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"tag":{"type":"string"},"email":{"type":"string"},"displayName":{"type":"string"},"status":{"type":"string"},"wonRound":{"type":"boolean"},"placement":{"type":"integer","format":"int32"},"winCount":{"type":"integer","format":"int32"},"lossCount":{"type":"integer","format":"int32"},"eventsAttended":{"type":"integer","format":"int32"},"winLossRatio":{"type":"number","format":"double"}}},"DTOBoolean":{"type":"object","properties":{"statusCode":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"boolean"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Jwt":{"type":"object","properties":{"tokenValue":{"type":"string"},"issuedAt":{"type":"integer","format":"int64"},"expiresAt":{"type":"integer","format":"int64"},"headers":{"type":"object","additionalProperties":{"type":"object"}},"claims":{"type":"object","additionalProperties":{"type":"object"}},"id":{"type":"string"},"notBefore":{"type":"integer","format":"int64"},"subject":{"type":"string"},"audience":{"type":"array","items":{"type":"string"}},"issuer":{"type":"string","format":"url"}}},"JwtAuthenticationToken":{"type":"object","properties":{"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"details":{"type":"object"},"authenticated":{"type":"boolean"},"principal":{"type":"object"},"credentials":{"type":"object"},"token":{"$ref":"#/components/schemas/Jwt"},"name":{"type":"string"},"tokenAttributes":{"type":"object","additionalProperties":{"type":"object"}}}},"Bracket":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"tournament":{"$ref":"#/components/schemas/Tournament"},"rounds":{"type":"array","items":{"$ref":"#/components/schemas/Round"}}}},"DTOListTournament":{"type":"object","properties":{"statusCode":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Tournament"}}}},"Round":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"entrants":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}},"Tournament":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"capacity":{"type":"integer","format":"int32"},"points":{"type":"integer","format":"int32"},"gameTitle":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"imgUrl":{"type":"string"},"startDate":{"type":"string","format":"date-time"},"isActive":{"type":"boolean"},"owner":{"$ref":"#/components/schemas/User"},"bracketType":{"type":"string"},"stageList":{"type":"array","items":{"type":"string"}},"bracket":{"$ref":"#/components/schemas/Bracket"}}},"DTOTournament":{"type":"object","properties":{"statusCode":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/Tournament"}}},"DTOString":{"type":"object","properties":{"statusCode":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"string"}}},"DTOListUser":{"type":"object","properties":{"statusCode":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}